backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-docx==0.8.11
pdfplumber==0.7.5
PyPDF2==3.1.0
requests==2.31.0
python-dotenv==1.0.0
mammoth==1.10.0
